Fixing three failing tests / remove duplicate.
Three of the tests were failing because there was a hard-wired expiration date of 2012 that was causing the is_valid() check to fail.

I've set up the tests to use a relative delta so that the tests will always pass no matter what your system date is.

One of the tests was also a copy-paste error and has been removed.

All tests are passing now. On to some TDD on the Authorize library.